NCS-477 Peding refactor did not get committed
