+++ src/main/java/uk/gov/companieshouse/confirmationstatementapi/model/dao/ConfirmationStatementSubmissionDataDao.java
@@ -1,6 +1,7 @@
 package uk.gov.companieshouse.confirmationstatementapi.model.dao;
 
 import org.springframework.data.mongodb.core.mapping.Field;
+import uk.gov.companieshouse.confirmationstatementapi.model.dao.statementofcapital.StatementOfCapitalDataDao;
 
 public class ConfirmationStatementSubmissionDataDao {
     @Field("statement_of_capital_data")
